Ethan Dastick (2024-12-10) Merged Packet_Decomposition_and_IP_reading back into main.
Ethan Dastick (2024-12-02)
Added getDomain function in place of getIP. Attempts to use nslookup to get domain names from IP so we can block the domain.

Created checkBlacklist function. Compares the requested source IP against a list of blacklisted ips to determine if it is malicious. If so, calls getDomain and then BlockDomain.