Ethan Dastick (2024-12-10) Updated Log function - it now takes in multiple parameters and outputs them all to a csv file in multiple columns.
Main has been fleshed out - the software now has a basic running structure. Several key aspects are still not complete, such as reading from the terminal (necessary for nslookup and capturePackets) but everything else is more or less operational. Minor bugs remain in the log function but basic functionality is working.

Sha256 algorithm is not used but operational